acquire_lock_ldlstcを実装した．

+++ sample/os/atk2-sc1-mc_1.4.2/arch/v850_gcc/prc_ccb.h
@@ -68,7 +68,7 @@
 /*
  *  ロックの型
  */
-typedef uint32 LockType;
+typedef uint16 LockType;
 
 /*
  *  前方参照
